# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project WebAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sic_code": {
            "get": {
                "tags": [
                    "Helper"
                ],
                "summary": "List of sic code",
                "description": "List of sic code",
                "operationId": "list_sic_code",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/state": {
            "get": {
                "tags": [
                    "Helper"
                ],
                "summary": "List of state",
                "description": "List of state",
                "operationId": "list_state",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}